pyproject.toml
[tool.poetry]
name = "mindsolver-fastapi"
version = "0.1.0"
description = ""
authors = ["NamisMe <nam_ing@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
python-dotenv = "^1.0.1"
openai = "^1.9.0"
uvicorn = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
